requirements.txt
git+https://github.com/video-db/videodb-python@experimental
openai>=1.30.0
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0.0
langgraph>=0.2.0
aiohttp>=3.9.0

# Object detection (required for local detection mode)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.40.0
Pillow>=10.0.0
tqdm>=4.66.0
